requirements.txt
yt-dlp
openai-whisper
torch
numpy
minio
temporalio
requests
